+++ infrastructure/tests/factories.py
@@ -1,11 +1,9 @@
 import factory
 
-from .models import Project, Initiative
-
 
 class ProjectFactory(factory.django.DjangoModelFactory):
     class Meta:
-        model = Project
+        model = 'infrastructure.Project'
 
     name = factory.Sequence(lambda n: 'Project %d' % n)
     slug = factory.Sequence(lambda n: 'Project-%d' % n)
@@ -13,7 +11,7 @@ class ProjectFactory(factory.django.DjangoModelFactory):
 
 class InitiativeFactory(factory.django.DjangoModelFactory):
     class Meta:
-        model = Initiative
+        model = 'infrastructure.Initiative'
 
     name = factory.Sequence(lambda n: 'Initiative %d' % n)
     slug = factory.Sequence(lambda n: 'Initiative-%d' % n)
